+++ webpack/containers/questions/QuestionItem.js
@@ -2,15 +2,15 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { questionProps } from "../prop-types/question_props";
-import { responseSetProps } from "../prop-types/response_set_props";
-import currentUserProps from '../prop-types/current_user_props';
-import { surveillanceSystemsProps }from '../prop-types/surveillance_system_props';
-import { surveillanceProgramsProps } from '../prop-types/surveillance_program_props';
-import { fetchSearchResults } from '../actions/search_results_actions';
-import SearchResult from '../components/SearchResult';
-import DashboardSearch from '../components/DashboardSearch';
-import SearchResultList from '../components/SearchResultList';
+import { questionProps } from "../../prop-types/question_props";
+import { responseSetProps } from "../../prop-types/response_set_props";
+import currentUserProps from '../../prop-types/current_user_props';
+import { surveillanceSystemsProps }from '../../prop-types/surveillance_system_props';
+import { surveillanceProgramsProps } from '../../prop-types/surveillance_program_props';
+import { fetchSearchResults } from '../../actions/search_results_actions';
+import SearchResult from '../../components/SearchResult';
+import DashboardSearch from '../../components/DashboardSearch';
+import SearchResultList from '../../components/SearchResultList';
 import { Modal, Button } from 'react-bootstrap';
 
 const QUESTION_ITEM_CONTEXT = 'QUESTION_ITEM_CONTEXT';
