+++ webpack/landing.js
@@ -10,9 +10,9 @@ import { hashHistory, Router, Route, IndexRoute } from 'react-router';
 import DashboardContainer from './containers/DashboardContainer';
 import ResponseSetShowContainer from './containers/response_sets/ResponseSetShowContainer';
 import FormShowContainer from './containers/FormShowContainer';
-import QuestionShowContainer from './containers/QuestionShowContainer';
+import QuestionShowContainer from './containers/questions/QuestionShowContainer';
 import ResponseSetEditContainer from './containers/response_sets/ResponseSetEditContainer';
-import QuestionEditContainer from './containers/QuestionEditContainer';
+import QuestionEditContainer from './containers/questions/QuestionEditContainer';
 import SurveyEditContainer from './containers/SurveyEditContainer';
 import FormsEditContainer from './containers/FormsEditContainer';
 import SurveyShowContainer from './containers/surveys/ShowContainer';
